move parameterize.ts
