+++ packages/nimble-components/src/anchor-tabs/tests/anchor-tabs.spec.ts
@@ -13,7 +13,7 @@ import '../../anchor-tab';
 import type { AnchorTab } from '../../anchor-tab';
 import { waitForUpdatesAsync } from '../../testing/async-helpers';
 import { fixture, Fixture } from '../../utilities/tests/fixture';
-import { parameterizeNamedList } from '../../utilities/tests/parameterized';
+import { parameterizeNamedList } from '../../testing/parameterized';
 
 describe('AnchorTabs', () => {
     let element: AnchorTabs;
