+++ packages/nimble-components/src/label-provider/table/tests/label-provider-table.spec.ts
@@ -2,7 +2,7 @@ import { spinalCase } from '@microsoft/fast-web-utilities';
 import { html } from '@microsoft/fast-element';
 import * as labelTokensNamespace from '../label-tokens';
 import { LabelProviderTable, labelProviderTableTag } from '..';
-import { parameterizeNamedList } from '../../../utilities/tests/parameterized';
+import { parameterizeNamedList } from '../../../testing/parameterized';
 import {
     getAttributeName,
     getPropertyName,
