+++ packages/nimble-components/src/rich-text-mention/users/view/tests/rich-text-mention-users-view.spec.ts
@@ -3,7 +3,7 @@ import { RichTextMentionUsersView, richTextMentionUsersViewTag } from '..';
 import { type Fixture, fixture } from '../../../../utilities/tests/fixture';
 import { waitForUpdatesAsync } from '../../../../testing/async-helpers';
 import { RichTextMentionUsersViewPageObject } from '../testing/rich-text-mention-users-view.pageobject';
-import { parameterizeNamedList } from '../../../../utilities/tests/parameterized';
+import { parameterizeNamedList } from '../../../../testing/parameterized';
 import { wackyStrings } from '../../../../utilities/tests/wacky-strings';
 
 async function setup(): Promise<Fixture<RichTextMentionUsersView>> {
