+++ packages/nimble-components/src/table-column/duration-text/models/tests/duration-formatter.spec.ts
@@ -1,4 +1,4 @@
-import { parameterizeNamedList } from '../../../../utilities/tests/parameterized';
+import { parameterizeNamedList } from '../../../../testing/parameterized';
 import { DurationFormatter } from '../duration-formatter';
 
 describe('Duration Formatter', () => {
