+++ packages/nimble-components/src/table-column/number-text/models/tests/default-formatter.spec.ts
@@ -1,4 +1,4 @@
-import { parameterizeNamedList } from '../../../../utilities/tests/parameterized';
+import { parameterizeNamedList } from '../../../../testing/parameterized';
 import { DefaultFormatter } from '../default-formatter';
 
 describe('DefaultFormatter', () => {
