+++ packages/nimble-components/src/text-area/tests/text-area.spec.ts
@@ -2,7 +2,7 @@ import { html } from '@microsoft/fast-element';
 import { TextArea, textAreaTag } from '..';
 import { waitForUpdatesAsync } from '../../testing/async-helpers';
 import { fixture, Fixture } from '../../utilities/tests/fixture';
-import { parameterizeNamedList } from '../../utilities/tests/parameterized';
+import { parameterizeNamedList } from '../../testing/parameterized';
 
 async function setup(): Promise<Fixture<TextArea>> {
     return fixture<TextArea>(html`<nimble-text-area></nimble-text-area>`);
