+++ packages/nimble-components/src/utilities/tests/tests/parametized.spec.ts
@@ -1,4 +1,4 @@
-import { parameterize, parameterizeNamedList } from '../parameterized';
+import { parameterize, parameterizeNamedList } from '../../../testing/parameterized';
 
 // The following aliases are just to reduce the number
 // of eslint disables in this test file. In normal
